Refactor Register/Login components for improved styling and responsiveness
