+++ client/src/components/layout/Navbar.module.css
@@ -10,6 +10,10 @@
   z-index: 10;
 }
 
+.navbar.bgTransparent {
+  background: none;
+}
+
 .brand {
   font-size: 14px;
   display: flex;
