--- client/src/pages/Register.module.css
@@ -1,228 +1,0 @@
-.registerPage {
-  margin: 0;
-  padding: 0;
-  width: 100%;
-  height: 100%;
-  overflow: hidden; /* Prevent scrolling */
-  font-family: 'Inter', sans-serif;
-  box-sizing: border-box;
-}
-
-/* Background SVG */
-.waveBackground {
-  position: fixed; /* Use fixed to cover the entire viewport */
-  top: 0;
-  left: 0;
-  width: 100vw; /* Use viewport width */
-  height: 100vh; /* Use viewport height */
-  z-index: -1; /* Ensure it stays behind other content */
-  animation: float 5s ease-in-out infinite;
-}
-
-/* Floating animation */
-@keyframes float {
-  0%,
-  100% {
-    transform: translateY(0);
-  }
-  50% {
-    transform: translateY(-20px);
-  }
-}
-
-/* Overall container layout */
-.registerContainer {
-  display: flex;
-  height: 100vh;
-  justify-content: flex-start; /* Move everything to the right */
-  align-items: center;
-  padding-left: 140px; /*Adjust to move further right */
-}
-
-/* Left Side - Welcome Section */
-.welcomeSection {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: left;
-  background-color: transparent;
-  height: 100%;
-  padding: 1rem;
-  margin-left: 80px;
-  max-width: 400px;
-}
-
-.welcomeTitle {
-  display: flex;
-  align-items: left;
-  font-size: 45px;
-  font-weight: 800;
-  line-height: 50px;
-  letter-spacing: -0.45px;
-  color: var(--light);
-}
-
-.welcomeText {
-  font-size: 18px;
-  font-weight: 400;
-  letter-spacing: 1px;
-  text-align: left;
-  color: var(--light);
-  max-width: 400px;
-}
-
-/* Right Side - Form Section */
-.formSection {
-  flex: 1;
-  background: transparent;
-  padding: 2rem;
-  border-radius: 7px;
-  text-align: center;
-  width: 100%; /* Ensures it takes full width */
-  max-width: 310px; /* Limit max width for better design */
-  margin: 0 auto; /* Center the form section */
-  margin-top: 80px; /* Move the form down */
-}
-
-.formTitle {
-  text-align: center;
-  margin-bottom: 1rem;
-  color: #333;
-  font-size: 24px;
-  font-weight: 700;
-}
-
-/* Form Layout */
-.registerForm {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-}
-
-/* Input Wrapper */
-.inputWrapper {
-  display: flex;
-  align-items: center;
-  border: 1px solid #ccc;
-  border-radius: 15px;
-  padding: 0.5rem;
-  background-color: var(--neutral);
-  position: relative;
-  width: 100%; /* Full width of the form section */
-  box-sizing: border-box; /* Include padding and border in width */
-  max-width: 300px; /* Limit max width to avoid too wide fields */
-  overflow: hidden; /* Ensure the container doesn't overflow */
-}
-
-/* Input Field */
-.inputWrapper input {
-  flex: 1;
-  border: none;
-  outline: none;
-  font-size: 1rem;
-  padding: 0.5rem 2.5rem 0.5rem 0.5rem; /* Add padding to the right for the icon */
-  background-color: var(--neutral);
-  box-sizing: border-box;
-  width: 100%; /* Ensure input takes full width */
-  overflow: hidden; /* Hide overflow */
-  text-overflow: ellipsis; /* Add ellipsis for overflow text */
-  white-space: nowrap; /* Prevent text from wrapping */
-}
-
-/* Style for input placeholder text */
-.inputWrapper input::placeholder {
-  color: var(--primary);
-  opacity: 1;
-  font-size: 0.8rem;
-}
-
-/* Icon Inside Input */
-.inputWrapper img {
-  width: 20px;
-  height: 20px;
-  position: absolute;
-  right: 10px;
-  top: 50%;
-  transform: translateY(-50%);
-}
-
-.eyeIcon {
-  pointer-events: auto; /* Allow clicks */
-  cursor: pointer; /* Show clickable cursor */
-  position: absolute;
-  right: 10px;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 20px;
-  height: 20px;
-}
-
-/* Error Messages */
-.errorMessage {
-  color: red;
-  font-size: 0.7rem;
-  text-align: left;
-  margin-top: -0.5rem;
-}
-
-/* Login Redirect */
-.loginRedirect {
-  margin-top: 0.1rem;
-  font-size: 0.8rem;
-  text-align: left;
-}
-
-.loginLink {
-  color: #007bff;
-  font-weight: 600;
-  cursor: pointer;
-  text-decoration: none;
-}
-
-.loginLink:hover {
-  text-decoration: underline;
-}
-
-/* Register Button */
-.registerForm button {
-  width: 70%; /* Smaller than input fields */
-  max-width: 150px; /* Ensures it doesn't get too wide */
-  margin: 0 auto; /* Centers the button */
-  display: block;
-  text-align: center;
-  border-radius: 15px;
-  padding: 0.5rem 1rem;
-  font-size: 1rem;
-  cursor: pointer;
-}
-
-/* Responsive Design */
-@media (max-width: 600px) {
-  .formSection {
-    width: 90%; /* Take more space on small screens */
-    padding: 1rem;
-  }
-
-  @media (max-width: 480px) {
-    .registerContainer {
-      flex-direction: column;
-      align-items: center;
-      padding: 1.5rem;
-    }
-  }
-
-  .inputWrapper {
-    max-width: 100%; /* Allow full width on small screens */
-  }
-}
-
-/* Override browser autofill styles */
-.inputWrapper input:-webkit-autofill,
-.inputWrapper input:-webkit-autofill:hover,
-.inputWrapper input:-webkit-autofill:focus,
-.inputWrapper input:-webkit-autofill:active {
-  -webkit-box-shadow: 0 0 0 1000px var(--neutral) inset !important; /* Match your input background color */
-  -webkit-text-fill-color: var(--dark) !important; /* Match your text color */
-  transition: background-color 5000s ease-in-out 0s; /* Prevent background color change */
-}
